requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
python-chess
stockfish
werkzeug